Update 0,1,3
Hi Jeff and Akmal,

could you please post the following on the web page asap:

0. Replace red box with deadline extension with blue box that says:
Paper submission is now closed.

1. preliminary program (you can do it as a pdf file)

2. student travel form (word document or fillable pdf)

3. Paste the following text under the appropriate headings:

Student Travel Grants
Graduate students who are presenting a paper or participating in the
demo competition are welcome to apply for a travel grant of $500 by
March 30th. To apply, please fill out this form: (link to form) and
email to wptc2015@colorado.edu. Notification of grant selections will be
announced on April 10, 2015. The selection criteria will include paper
ranking, demo proposal quality, and number of students from the same
institution who have applied. The student has to be registered for the
conference in order to receive a travel grant.

Hardware Demo Competition
A student demo competition will be held on Thursday, May 14, 2015, as a
part of WPTC2015, at the same location as the poster session.
Competitors need to send the following information by March 30, 2014, to
wptc2015@colorado.edu:
1.       One-page abstract of the demo, including a detailed block
diagram

2.       Names of student(s) competing; up to 3 students are allowed to
participate in each demo

3.       Name and institution of responsible faculty advisor

4.       Institution(s): student teams from more than one institutions
are allowed

5.       Necessary equipment (list everything you need)

If included in the list in item #5 above, the following can be
provided*:

-          Outlets and extension cords (110V, bring adaptors if needed)

-          Common power supplies

-          Low-frequency function generators and oscilloscopes

-          Multimeters

* For RF harvesting demos, please use cell phones or bring your own
FCC-compliant source. We may be able to provide one low-power density
source, but there is no guarantee that it will fit your demo.